Revert "move specific values files out of helm chart"
This reverts commit f1d290d0c8e17ab7412a8ef2cccb6e30614c9183.
The original commit, while improving the situation with the
Helmcharts, breaks the build pipeline.

The build pipeline expects the value files to be present in
the same directory as the charts, but they were moved out in
order to facilitate packing the web-ui helmchart for a custom
helm chart repo. I'm hoping to restore this commit on a later
date.